requirements.txt
requests
bs4
openai
google-genai
python-dotenv
tqdm